requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
SQLAlchemy==1.4.31
Flask-SQLAlchemy==2.5.1
celery==5.2.3
redis==4.1.4
moviepy==1.0.3
grpcio==1.66.1
grpcio-tools==1.66.1
python-dotenv==0.19.2
gunicorn==20.1.0
protobuf==5.28.2
Flask-Migrate